store/tikv, util: move storeutil to store/tikv
